+++ zarrs_metadata/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `ExtensionType` marker trait and `ExtensionType{DataType,ChunkGrid,ChunkKeyEncoding,Codec,StorageTranformer}`
 - Add `{Array,Group,Node}Metadata::to_string_pretty()`
 - Add `{Array,Group}Metadata{V2,V3}::to_string_pretty()`
+- Add `zlib` codec metadata
 
 ### Changed
 - **Breaking**: Move all codecs into a new `codec` module rather than the `v2`/`v3` modules